Add a new tool for VHFFS, it allows synching file-systems over networks in best-effort !
Designed for downloads repositories it doesn't currently scale a lot if 
files are oftenly modified.

It doesn't support uid/gid/mode/acl and will probably never do.

It doesn't support anything else than regular files, directories and 
symlink, and will probably never support the others.

It uses inotify(7), as a result it scales a lot even with billions of files 
and directories.

This is a monoprocess/monothread software featuring non-blocking I/O and 
zerocopy ways, so it reduces to the maximum possible the number of 
context-switchs.

Events are permanently scheduled (prioritized, ordered) based on what is 
happening.

The network layer scales to tens-gigabits per second.

It is just a great piece of software for synching slaves downloads 
servers (mirrors) from a hidden master server.