Fixed exim4 example config for PostgreSQL 9.x
PostgreSQL from 8.x to 9.x changed the PostgreSQL default setting of 
standard_conforming_strings from 'no' to 'yes', which means that the 
database now treats backslash characters in a string as ordinary 
characters. This in turn breaks the Exim implementation of 
pgsql_quote(), which apparently adds a backslash in front of every 
percent and underscore character, leading to corrupted SQL statements 
being sent to the SQL server.